0.9.11 - added option to set open and read timeouts for API calls
Fixes #8 and now enables people to set the open and read timeouts when required.

The default was 60 seconds, so I'm guessing this is to do with bad network connectivity.
Now in the case of a timeout it will error out and wait `poll_time` seconds before
contacting the API again.